backend/apps/api/auth/v1/serializers.py
from rest_framework_simplejwt.serializers import (
    TokenObtainPairSerializer as BaseTokenObtainPairSerializer,
)


class TokenObtainPairSerializer(BaseTokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        token["username"] = user.username
        token["isSuperUser"] = user.is_superuser
        token["isStaff"] = user.is_staff

        return token
